Changed - explicit casting of vector.size() for marshaling

+++ NaN2/include/module/team_module/snapshot.h
@@ -5,6 +5,7 @@
 #include <ProudNetServer.h>
 
 #include "../../world/player.h"
+#include <stdint.h>
 #include <vector>
 
 namespace nan2 {
@@ -36,10 +37,10 @@ inline CMessage& operator >> (CMessage& a, nan2::module::team_module::Snapshot& 
 }
 
 inline CMessage& operator << (CMessage& a, const nan2::module::team_module::Snapshot& packet) {
-  a << packet.team_snapshots.size();
+  a << (int32_t)packet.team_snapshots.size();
   for (auto& team_snapshot : packet.team_snapshots) {
     a << team_snapshot.id;
-    a << team_snapshot.players.size();
+    a << (int32_t)team_snapshot.players.size();
     for (auto& player : team_snapshot.players)
       a << player;
     a << team_snapshot.score;
